[UPD] Merge open_commercial_member_entity at parent addon level
